change error to warning for already-saved files
previously: if file existed and user did not set flag
`overwrite=True`, save to filesystem to would raise
an exception. a dictionary cataloging the output of the
save would therefore be returned only when a write
operation occurred

new behavior: if file exists and user does not set
flag `overwrite=True`, save to filesystem raises
a warning. in this case, no save is performed, but
the usual dictionary cataloging the output [of what
the save operation would have been] is returned
even when no write is performed, in addition to
the previous return behavior when a write operation
occurs.

food for thought: when saving a large database of items,
this may generate a lot of warnings and be annoying.
consider removing the warning, or allowing some kind
of verbosity setting to exist for databases/filesystems
to allow the warning message to be suppressed when
user is OK with the process flow.